requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
asyncio
certifi==2024.8.30
charset-normalizer==3.3.2
fastapi==0.115.0
gradio==4.44.0
httpx==0.27.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.5
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
nest-asyncio==1.6.0
openai==1.47.0
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
starlette==0.38.6
uvicorn==0.30.6
